Crate problem060

Problem 60 (Prime pair sets)

Solution detail

Use a standard depth-first backtracking algorithm, where each branch in the search tree represents adding a single prime to the set.

To speed up the search, branches can be pruned early when we know that the resultant sum cannot be smaller than that of some previously-found set of primes.

Finally, we can speed up checking by pre-calculating which pairs of primes concatenate to form another prime, rather than having to check each pair multiple times as we go through the search.

Structs

PrimePairTree
A structure that can be used for conducting a depth-first search for prime-pair sets.

PrimePairTreeStep
A description of a step that can be taken in the search tree.

PrimePairs
A structure that holds, for each prime, the primes that are able to concatenate with it, in both orders, to form new primes.

PrimeSet
A structure representing a set of primes, and their sum.

Functions

answer
Solve the problem, returning the answer as a String

concat
The result of concatenating the numbers p and q.

main

solve
Find the smallest sum of a set of primes such that any pair of primes in the set concatenate, in both orders, to form another prime. Only considers numbers up to the given maximum prime.